refactor: move ProductFormModal to features
